Remove unsigned integers from interfaces

+++ include/od/core/debug.h
@@ -12,7 +12,7 @@
 #define OD_LOG_LEVEL_LAST 6
 #define OD_LOG_LEVEL_DEFAULT OD_LOG_LEVEL_WARN
 
-#define OD_LOG_CONTEXT_CONSTRUCT() odLogContext_construct(__FILE__, __func__, __LINE__)
+#define OD_LOG_CONTEXT_CONSTRUCT() odLogContext_construct(__FILE__, __func__, static_cast<int32_t>(__LINE__))
 
 #if OD_BUILD_LOG
 #define OD_ERROR(...) odLog_log(OD_LOG_CONTEXT_CONSTRUCT(), OD_LOG_LEVEL_ERROR, __VA_ARGS__)
@@ -36,23 +36,23 @@
 struct odLogContext {
 	const char* file;
 	const char* function;
-	uint32_t line;
+	int32_t line;
 };
 
-OD_API_C OD_ENGINE_CORE_MODULE const char* odLogLevel_get_name(uint32_t log_level);
-OD_API_C OD_ENGINE_CORE_MODULE uint32_t odLogLevel_get_max(void);
-OD_API_C OD_ENGINE_CORE_MODULE void odLogLevel_set_max(uint32_t log_level);
+OD_API_C OD_ENGINE_CORE_MODULE const char* odLogLevel_get_name(int32_t log_level);
+OD_API_C OD_ENGINE_CORE_MODULE int32_t odLogLevel_get_max(void);
+OD_API_C OD_ENGINE_CORE_MODULE void odLogLevel_set_max(int32_t log_level);
 
 OD_API_C OD_ENGINE_CORE_MODULE struct odLogContext
-odLogContext_construct(const char* file, const char* function, uint32_t line);
+odLogContext_construct(const char* file, const char* function, int32_t line);
 
 OD_API_C OD_ENGINE_CORE_MODULE void
-odLog_log_variadic(struct odLogContext logger, uint32_t log_level, const char* format_c_str, va_list args);
+odLog_log_variadic(struct odLogContext logger, int32_t log_level, const char* format_c_str, va_list args);
 OD_API_C OD_ENGINE_CORE_MODULE void
-odLog_log(struct odLogContext logger, uint32_t log_level, const char* format_c_str, ...) OD_API_PRINTF(3, 4);
+odLog_log(struct odLogContext logger, int32_t log_level, const char* format_c_str, ...) OD_API_PRINTF(3, 4);
 OD_API_C OD_ENGINE_CORE_MODULE void
 odLog_assert(struct odLogContext logger, bool success, const char* expression_c_str);
 
-OD_API_C OD_ENGINE_CORE_MODULE char* odDebugString_allocate(uint32_t size);
+OD_API_C OD_ENGINE_CORE_MODULE char* odDebugString_allocate(int32_t size);
 OD_API_C OD_ENGINE_CORE_MODULE const char* odDebugString_format_variadic(const char* format_c_str, va_list args);
 OD_API_C OD_ENGINE_CORE_MODULE const char* odDebugString_format(const char* format_c_str, ...) OD_API_PRINTF(1, 2);
